feat: connect delete & get project list

+++ backend/app/api/project.py
@@ -1,8 +1,8 @@
-from flask import Blueprint, current_app
+from flask import Blueprint, current_app, request, json
 from flask_login import current_user, login_required
 from flask_restful import Api, Resource, abort, fields, marshal_with, reqparse
 
-from app.checker import check_create_project_param, check_project_permission
+from app.checker import check_create_project_param, check_project_permission, check_delete_project_password
 from app.service import ProjectService, UserService
 
 bp = Blueprint(
@@ -15,9 +15,10 @@ proj_service = ProjectService()
 user_service = UserService()
 
 parser = reqparse.RequestParser()
-parser.add_argument('creator_id', type=int, location='form')
+parser.add_argument('creator_name', type=str, location='form')
 parser.add_argument('project_name', type=str, location='form')
-parser.add_argument('project_language', type=str, location='form', choices=('python', 'cpp', 'typescript'), help='Bad choice: {error_msg}')
+parser.add_argument('language', type=str, location='form', choices=('Python', 'Cpp', 'Java', 'C'), help='Bad choice: {error_msg}')
+parser.add_argument('password', type=str, location='form')
 
 
 class Project(Resource):
@@ -68,15 +69,15 @@ class Project(Resource):
         tags:
             - Project
         parameters:
-            - name: creator_id
-              type: int
+            - name: creator_name
+              type: str
               required: true
               in: formData
             - name: project_name
               type: string
               required: true
               in: formData
-            - name: project_language
+            - name: language
               type: string
               required: true
               in: formData
@@ -88,13 +89,13 @@ class Project(Resource):
         key, flag = check_create_project_param(args)
         if not flag:
             abort(400, message="Invalid argument {}".format(key))
-        user, flag = user_service.find_user_by_id(args['creator_id'])
+        user, flag = user_service.find_user_by_username(args['creator_name'])
         if not flag:
-            abort(400, message="User {} not exist".format(args['creator_id']))
+            abort(400, message="User {} not exist".format(args['creator_name']))
         if current_user.username == user.username:
-            response, flag = proj_service.create_project(args['creator_id'], args['project_name'], args['project_language'])
+            response, flag = proj_service.create_project(args['creator_name'], args['project_name'], args['language'])
             if flag:
-                return response, 200
+                return response, 201
             else:
                 abort(400, message=response)
         else:
@@ -117,19 +118,27 @@ class Project(Resource):
               type: int
               required: true
               in: query
+            - name: password
+              type: string
+              required: true
+              in: data
         responses:
             204:
                 description: project delete success
             400:
                 description: no permission
+            401:
+                description: invalid password
             404:
                 description: project doesn't exist
             500:
-                description: delete project failed
+                description: internal error
 
         """
         if not check_project_permission(proj_id, "admin"):
             abort(400, message="Permission denied")
+        if not check_delete_project_password(current_user.username, json.loads(request.data)['password']):
+            abort(401, message="invalid password")
         proj, flag = proj_service.get_project(proj_id)
         if flag:
             res, flag = proj_service.remove_project(proj_id)
